reorganize number utils
