+++ e2e/utils/operators.ts
@@ -1,6 +1,6 @@
 import { Locator, Page } from 'playwright-core';
-import { prettifyNumber as pn } from '../../src/utils/helpers/prettifyNumber';
-export { prettifyNumber as pn } from '../../src/utils/helpers/prettifyNumber';
+import { prettifyNumber as pn } from '../../src/utils/helpers/number';
+export { prettifyNumber as pn } from '../../src/utils/helpers/number';
 
 export const isCI = !!process.env.CI && process.env.CI !== 'false';
 export const isDraft = !!process.env.DRAFT && process.env.DRAFT !== 'false';
