+++ src/components/strategies/create/BuySellBlock/useStrategyEvents.ts
@@ -8,7 +8,7 @@ import {
   StrategyEventTypeBase,
   StrategySellEventType,
 } from 'services/events/types';
-import { sanitizeNumberInput } from 'utils/helpers';
+import { sanitizeNumber } from 'utils/helpers';
 import { StrategyCreateSearch } from '../types';
 import { OrderCreate } from '../useOrder';
 import { useSearch } from 'libs/routing';
@@ -43,9 +43,9 @@ export const useStrategyEvents = ({
         buyOrderType: order.isRange ? 'range' : 'limit',
         buyBudget: order.budget,
         buyBudgetUsd: fiatValueUsd,
-        buyTokenPrice: sanitizeNumberInput(order.price, 18),
-        buyTokenPriceMin: sanitizeNumberInput(order.min, 18),
-        buyTokenPriceMax: sanitizeNumberInput(order.max, 18),
+        buyTokenPrice: sanitizeNumber(order.price, 18),
+        buyTokenPriceMin: sanitizeNumber(order.min, 18),
+        buyTokenPriceMax: sanitizeNumber(order.max, 18),
         strategyDirection: search?.strategyDirection,
         strategySettings: search?.strategySettings,
         strategyType: search?.strategyType,
@@ -57,9 +57,9 @@ export const useStrategyEvents = ({
       sellOrderType: order.isRange ? 'range' : 'limit',
       sellBudget: order.budget,
       sellBudgetUsd: fiatValueUsd,
-      sellTokenPrice: sanitizeNumberInput(order.price, 18),
-      sellTokenPriceMin: sanitizeNumberInput(order.min, 18),
-      sellTokenPriceMax: sanitizeNumberInput(order.max, 18),
+      sellTokenPrice: sanitizeNumber(order.price, 18),
+      sellTokenPriceMin: sanitizeNumber(order.min, 18),
+      sellTokenPriceMax: sanitizeNumber(order.max, 18),
       strategyDirection: search?.strategyDirection,
       strategySettings: search?.strategySettings,
       strategyType: search?.strategyType,
