+++ src/components/strategies/edit/tooltip/TooltipTokenRange.tsx
@@ -2,7 +2,7 @@ import { Tooltip } from 'components/common/tooltip/Tooltip';
 import { TokenPrice } from 'components/strategies/overview/strategyBlock/TokenPrice';
 import { Token } from 'libs/tokens';
 import { FC } from 'react';
-import { cn, prettifyNumber, sanitizeNumberInput } from 'utils/helpers';
+import { cn, prettifyNumber, sanitizeNumber } from 'utils/helpers';
 import { useFiatValue } from 'hooks/useFiatValue';
 import { LogoImager } from 'components/common/imager/Imager';
 
@@ -23,8 +23,8 @@ export const TooltipTokenRange: FC<TooltipPriceProps> = ({
   const { min, max } = range;
   const minFiatPrice = useFiatValue({ price: min, token });
   const maxFiatPrice = useFiatValue({ price: max, token });
-  const minFullPrice = sanitizeNumberInput(min, token.decimals);
-  const maxFullPrice = sanitizeNumberInput(max, token.decimals);
+  const minFullPrice = sanitizeNumber(min, token.decimals);
+  const maxFullPrice = sanitizeNumber(max, token.decimals);
   return (
     <Tooltip
       element={
